requirements.txt
Django
django-auth-ldap
django-crispy-forms
psycopg-binary
python-ldap
psycopg
crispy-bootstrap4
django-autocomplete-light
django-bootstrap-datepicker-plus
gunicorn
django-extensions
Pillow
django_rest_framework
requests
python-docx
docxtpl
docxcompose
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
crispy-bootstrap5
pymssql
